refactor components folder structure
