+++ frontend/pages/Error/Error.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import styles from 'styles/pages/Error.module.css';
+import styles from 'styles/pages/Error/Error.module.css';
 
 const Error = ({ title, message, buttonText, path }: any) => {
   const navigate = useNavigate();
