+++ frontend/pages/Questions/Questions.tsx
@@ -1,13 +1,13 @@
 import axios from 'api/axios';
-import CreateListSidepanel from 'components/CreateListSidepanel';
 import PaginationNavigation from 'components/pagination/PaginationNavigation';
-import QuestionListSelectable from 'components/QuestionListSelectable';
 import BasicSearch from 'components/search/BasicSearch';
 import useLocalStorage from 'hooks/useLocalStorage';
 import usePagination from 'hooks/usePagination';
+import CreateListSidepanel from 'pages/Questions/Component/CreateListSidepanel';
+import QuestionListSelectable from 'pages/Questions/Component/QuestionListSelectable';
 import React, { useEffect, useState } from 'react';
 import { useLocation, useSearchParams } from 'react-router-dom';
-import styles from 'styles/pages/Questions.module.css';
+import styles from 'styles/pages/Questions/Questions.module.css';
 import { useDebounce } from 'use-debounce';
 
 const Questions = () => {
