+++ frontend/src/App.tsx
@@ -3,11 +3,11 @@ import PersistLogin from 'components/auth/PersistLogin';
 import RequireAuth from 'components/auth/RequireAuth';
 import AppLayout from 'components/ui/AppLayout';
 import { AuthProvider } from 'context/AuthProvider';
-import List from 'pages/List';
-import Lists from 'pages/Lists';
-import Login from 'pages/Login';
-import NotFound from 'pages/NotFound';
-import Questions from 'pages/Questions';
+import NotFound from 'pages/Error/NotFound';
+import List from 'pages/List/List';
+import Lists from 'pages/Lists/Lists';
+import Login from 'pages/Login/Login';
+import Questions from 'pages/Questions/Questions';
 import React from 'react';
 import {
   createBrowserRouter,
